Replace path navigatations with @ alias
