+++ cypress/support/apiHelpers.ts
@@ -1,7 +1,7 @@
 import { faker } from "@faker-js/faker";
 
 // Function to create form-urlencoded body
-export const createFormUrlEncodedBody = (data) => {
+export const createFormUrlEncodedBody = (data: any) => {
   return Object.fromEntries(data);
 };
 
@@ -11,7 +11,7 @@ export const formUrlEncodedHeaders = {
 };
 
 // Function to generate a new user body
-export const generateUserBody = (existingBody) => {
+export const generateUserBody = (existingBody: any) => {
   const newBody = new Map(existingBody);
   newBody.set("title", faker.word.noun());
   return newBody;
