+++ cypress/support/common-functions.ts
@@ -1,0 +1,3 @@
+export function getRandomInt(min: number, max: number) {
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
