+++ cypress/tests/UI/pages/account-created-page.ts
@@ -4,10 +4,6 @@ class AccountCreatedPage {
     accountDeletedBackTxt: () => cy.get("b"),
     continueBtn: () => cy.get("[data-qa='continue-button']"),
   };
-
-  navigateToLoginPage() {
-    this.Selectors.createAccountFeedBackTxt().click();
-  }
 }
 
 export const accountCreatedPage = new AccountCreatedPage();
