--- cypress/utils/common-functions.ts
@@ -1,3 +1,0 @@
-export function getRandomInt(min, max) {
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-}
